# mir: remove the tree delimiter nodes