… compiler/mir/mirpasses.nim
… compiler/mir/rtchecks.nim
… compiler/mir/treechangesets.nim
… compiler/sem/aliasanalysis.nim
… compiler/sem/injectdestructors.nim
… compiler/sem/mirexec.nim
… compiler/vm/packed_env.nim
… compiler/vm/vmserialize.nim
… doc/mir.rst
… tests/compiler/tdatatables.nim
… tests/compiler/tmir_trees.nim
… tests/compiler/ttreechangesets.nim