## Summary

Instead of marking the end of a subtree with a dedicated node, the
number of child nodes in a subtree is now stored in the subtree's root
node. The idea is to:
* reduce the memory footprint of MIR trees
* make the storage format more streamlined
* simplify some traversal

## Details

The idea of using dedicated end nodes originated at a time where:
* fast back-/up-wards MIR traversal was required
* statements could be nested

Neither is the case anymore, and replacing the delimiter `mnkEnd` nodes
with storing the length has some benefits:
* significantly less nodes per `MirTree` (reduced memory usage, faster
  scanning, faster copying, etc.)
* `O(1)` length lookup
* lower cost of changing a subtree root node, or replacing it with its
  single child node
* simpler recursive traversal, as the end nodes don't have to be
  ignored/considered anymore

The downside of removing the delimiters nodes are that:
* slightly higher tree construction cost: the number of child nodes
  needs to be kept track of
* walking "upwards" (i.e., towards the root) in the tree has a higher
  cost

Ultimately, the upsides outweigh the downsides, especially since
walking the tree upwards is seldomly done.

### Adjustments

For memory efficiency, some subtree root nodes (like `mnkPathPos`)
stored extra information in the root node. Since all subtree root nodes
now need to store the length, this is no longer possible, and the
information has to be moved into separate nodes:
* for call trees, whether the call has side-effects is stored as an
  immediate value (using the new `mnkImmediate` node) in the tree's
  first slot
* for field/positional access, the field/position is stored in the
  tree's second slot

For easier processing of `mnkObjConstr`/`mnkRefConstr` trees, the
`mnkBinding` is introduced, which groups the `mnkField` and associated
expression together in a subtree.

The rest of the changes are about:
* updating the `MirBuilder` routines to track the child node count and
  patch the root nodes
* updating the construction of MIR trees in various places
* updating the various tree traversal logic that expected the old
  layout