@zerbina 2024-06-09
Using a `-d:release` compiler and compiling the current `devel` (96cefc4fd6d78ff5fd26555db3bf30d8257f8d5e) compiler, I get the following benchmarks results:
```
hyperfine --warmpup 1 nim c --compileOnly --verbosity:0 --hints:off --warnings:off compiler/nim.nim

before  15.218 s ±  0.181 s
now     14.989 s ±  0.169 s
```